WIP - add caching and related tests
